fix: correct merge

+++ packages/lethebot/tg_webhook/bot.py
@@ -44,22 +44,25 @@ class LetheBot:
             await self._send_message(
                 update.effective_chat.id,
                 f"Hi {user.mention_markdown()}! I am Lethe (Ле́та)!\nI can hide you from sensitive chats.",
+                update_alarm_message=False,
             )
             await self._send_message(
                 update.effective_chat.id,
                 f"Please send the following message to 2 people you trust.\nYou will be able to recover your chats if they tell me you are safe.",
+                update_alarm_message=False,
             )
             await self._send_message(
                 update.effective_chat.id,
                 helpers.escape_markdown(f"Dear friend, please press here\nhttps://t.me/{self.bot.username}?start={self.invite_code}."),
+                update_alarm_message=False,
             )
-            keyboard = [
+            keyboard = InlineKeyboardMarkup([
                 [
                     InlineKeyboardButton("😎 Yes", callback_data=json.dumps(
                         {'action': 'begin_review'}
                     )),
                 ]
-            ]
+            ])
             await self._send_message(
                 update.effective_chat.id,
                 f"Are you ready to review your chats?",
@@ -144,7 +147,7 @@ class LetheBot:
         reply_markup = InlineKeyboardMarkup(keyboard)
         await self._send_message(
             update.effective_chat.id,
-            f'Is this a sensitive chat?\n💬{selected_chat.title or selected_chat.id}\n\n{self.generate_progress_bar(marked_chats, total_chats)}',
+            helpers.escape_markdown(f'Is this a sensitive chat?\n💬{selected_chat.title or selected_chat.id}\n\n{self.generate_progress_bar(marked_chats, total_chats)}'),
             reply_markup=reply_markup
         )
 
@@ -154,11 +157,6 @@ class LetheBot:
 
 
     async def button_yesno(self, update: Update, query: CallbackQuery, callback_data: dict):
-        if callback_data['action'] == 'begin_review':
-            # TODO move begin_review and other actions to Enum
-            await self.get_chat(update)
-            return
-
         data = await self.tg_client._read_saved_message()
         if callback_data['action'] == 'yes':
             data.get('chats')[callback_data['chat_id']] = {
@@ -216,13 +214,12 @@ class LetheBot:
             action = callback_data['action']
             if action == 'yes' or action == 'no':
                 return await self.button_yesno(update, query, callback_data)
+            elif action == 'begin_review':
+                return await self.get_chat(update)
             elif action == 'safe':
                 return # TODO safe vote
         elif update.message.text.startswith('/start'):
             return await self.start(update)
-
-        if update.message.text.startswith('/start'):
-            return await self.start(update, context)
         elif update.message.text == '/get_chat':
             return await self.get_chat(update)
         elif update.message.text == '/clear':
@@ -236,30 +233,34 @@ class LetheBot:
         print(f'unhandled update {update}')
 
 
-    async def _send_message(self, chat_id:int, text:str, reply_markup:InlineKeyboardMarkup=None):
+    async def _send_message(self, chat_id:int, text:str, reply_markup:InlineKeyboardMarkup=None, update_alarm_message=True):
         data = await self.tg_client._read_saved_message()
-        if 'alarm_message_id' in data:
-            await self.bot.delete_message(chat_id=chat_id, message_id=data['alarm_message_id'])
+        if 'alarm_message_id' in data and update_alarm_message:
+            try:
+                await self.bot.delete_message(chat_id=chat_id, message_id=data['alarm_message_id'])
+            except BaseException as e:
+                pass
         await self.bot.send_message(
             chat_id=chat_id,
             text=text,
             reply_markup=reply_markup,
             parse_mode=constants.ParseMode.MARKDOWN,
         )
-        keyboard = [
-            [
-                InlineKeyboardButton("🚨SOS🚨", callback_data=json.dumps({'action': 'sos'}))
+        if update_alarm_message:
+            keyboard = [
+                [
+                    InlineKeyboardButton("🚨SOS🚨", callback_data=json.dumps({'action': 'sos'}))
+                ]
             ]
-        ]
-        reply_markup = InlineKeyboardMarkup(keyboard)
-        alarm_msg = await self.bot.send_message(
-            chat_id=chat_id,
-            text=f'🚨🚨🚨🚨🚨\n\n\nВ ЭКСТРЕННОЙ СИТУАЦИИ\nНАЖМИТЕ SOS\n\n\n🚨🚨🚨🚨🚨',
-            reply_markup=reply_markup,
-            parse_mode=constants.ParseMode.MARKDOWN,
-        )
-        data['alarm_message_id'] = alarm_msg.id
-        await self.tg_client._write_saved_message(data)
+            reply_markup = InlineKeyboardMarkup(keyboard)
+            alarm_msg = await self.bot.send_message(
+                chat_id=chat_id,
+                text=f'🚨🚨🚨🚨🚨\n\n\nВ ЭКСТРЕННОЙ СИТУАЦИИ\nНАЖМИТЕ SOS\n\n\n🚨🚨🚨🚨🚨',
+                reply_markup=reply_markup,
+                parse_mode=constants.ParseMode.MARKDOWN,
+            )
+            data['alarm_message_id'] = alarm_msg.id
+            await self.tg_client._write_saved_message(data)
 
     def get_bot(self, token: str):
         application = (
